Verifica si cliente existe antes de crear una cuenta.

+++ src/main/java/com/devsu/challenge/backend/apirest/accountapp/controller/CuentaController.java
@@ -1,9 +1,12 @@
 package com.devsu.challenge.backend.apirest.accountapp.controller;
 
 import com.devsu.challenge.backend.apirest.accountapp.entity.Cuenta;
+import com.devsu.challenge.backend.apirest.accountapp.exceptions.ClienteNoEncontradoException;
 import com.devsu.challenge.backend.apirest.accountapp.service.ICuentaService;
 import java.util.List;
 import java.util.Optional;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -14,7 +17,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/cuentas")
+@RequestMapping("/cuentas")
 public class CuentaController {
 
     private final ICuentaService cuentaService;
@@ -24,27 +27,52 @@ public class CuentaController {
     }
 
     @GetMapping
-    public List<Cuenta> getAllCuentas() {
-        return cuentaService.getAllCuentas();
+    public ResponseEntity<List<Cuenta>> getAllCuentas() {
+        List<Cuenta> cuentas = cuentaService.getAllCuentas();
+        if (cuentas.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(cuentas, HttpStatus.OK);
     }
 
     @PostMapping
-    public Cuenta createCuenta(@RequestBody Cuenta cuenta) {
-        return cuentaService.createCuenta(cuenta);
+    public ResponseEntity<Cuenta> crearCuenta(@RequestBody Cuenta cuenta) {
+        Cuenta nuevaCuenta = cuentaService.crearCuenta(cuenta);
+        try {
+            return new ResponseEntity<>(nuevaCuenta, HttpStatus.CREATED);
+        } catch (ClienteNoEncontradoException cliEx) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @GetMapping("/{id}")
-    public Optional<Cuenta> getCuentaById(@PathVariable Long id) {
-        return cuentaService.getCuentaById(id);
+    public ResponseEntity<Cuenta> getCuentaById(@PathVariable Long id) {
+        Optional<Cuenta> cuenta = cuentaService.getCuentaById(id);
+        return cuenta
+            .map(cli -> new ResponseEntity<>(cli, HttpStatus.OK))
+            .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
     @PutMapping("/{id}")
-    public Cuenta updateCuenta(@PathVariable Long id, @RequestBody Cuenta cuentaDetails) {
-        return cuentaService.updateCuenta(id, cuentaDetails);
+    public ResponseEntity<Cuenta> updateCuenta(@PathVariable Long id, @RequestBody Cuenta cuentaDetalle) {
+        Cuenta updatedCuenta = cuentaService.updateCuenta(id, cuentaDetalle);
+        if (updatedCuenta == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(updatedCuenta, HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteCuenta(@PathVariable Long id) {
-        cuentaService.deleteCuenta(id);
+    public ResponseEntity<Void> deleteCuenta(@PathVariable Long id) {
+        Optional<Cuenta> cuenta = cuentaService.getCuentaById(id);
+
+        if (cuenta.isPresent()) {
+            cuentaService.deleteCuenta(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 }
